requirements.txt
python-fasthtml
python-dotenv
google-genai
uvicorn
gspread
google-auth
google-cloud-storage
pytz
markdown
